requirements.txt
langchain 
langchain-community
langchain-groq
langchainhub 
groq 
openai 
python-dotenv
streamlit
